summary_perf: remove sql to filter test projects from summaries to py functions

+++ src/api/utils/summaries.py
@@ -1,9 +1,12 @@
 from ..models import (
+    Project,
     SummarySampleEventModel,
     SummarySampleEventSQLModel,
     SummarySiteModel,
     SummarySiteSQLModel,
 )
+
+project_statuses = [Project.OPEN, Project.LOCKED]
 
 
 def update_project_summaries(project_id):
@@ -11,7 +14,13 @@ def update_project_summaries(project_id):
     update_project_summary_sample_event(project_id)
 
 
-def update_project_summary_site(project_id):
+def update_project_summary_site(project_id, skip_test_project=True):
+    if (
+        skip_test_project
+        and Project.objects.filter(pk=project_id, status=Project.TEST).exists()
+    ):
+        return
+
     SummarySiteModel.objects.filter(project_id=project_id).delete()
     for record in SummarySiteSQLModel.objects.all().sql_table(project_id=project_id):
         values = {
@@ -21,7 +30,13 @@ def update_project_summary_site(project_id):
         SummarySiteModel.objects.create(**values)
 
 
-def update_project_summary_sample_event(project_id):
+def update_project_summary_sample_event(project_id, skip_test_project=True):
+    if (
+        skip_test_project
+        and Project.objects.filter(pk=project_id, status=Project.TEST).exists()
+    ):
+        return
+
     SummarySampleEventModel.objects.filter(project_id=project_id).delete()
     for record in SummarySampleEventSQLModel.objects.all().sql_table(
         project_id=project_id
